refactor: move init_remote_index and InitResult to CheckRemoteIndexMixin
